% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coin_flip.R
\name{abc_coin_flip_game}
\alias{abc_coin_flip_game}
\title{Simulate many coin flip games via Approximate Bayesian Computation}
\usage{
abc_coin_flip_game(players, result = seq_along(players), iters = 10000,
  cores = 1L)
}
\arguments{
\item{players}{A list of players created with \code{\link{new_player}}.}

\item{result}{The final ranking of players in the game. The winner should
receive a 1, second place a 2, etc. Set \code{NULL} to leave the result
open (e.g., to compute win probabilities ahead of a game).}

\item{iters}{Number of samples to take from the joint distribution
\eqn{(d*, \theta*)}.}

\item{cores}{Number of cores to use for the computation. The special value
\code{cores = -1} uses all the cores (detected by
\code{parallel::detectCores}).}
}
\value{
A \code{bcf_game} object.
}
\description{
Simulate many coin flip games via Approximate Bayesian Computation
}
\examples{
\dontrun{
players <- list(new_player("Tom"), new_player("Bob"))
sim_game <- abc_coin_flip_game(players)
}
}
