% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game.R
\name{as.data.frame.bcf_game}
\alias{as.data.frame.bcf_game}
\title{S3 as.data.frame method for game}
\usage{
\method{as.data.frame}{bcf_game}(x, ...)
}
\arguments{
\item{x}{S3 game object.}

\item{...}{Additional parameters.}
}
\value{
A data frame.
}
\description{
S3 as.data.frame method for game
}
\keyword{internal}
