% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player.R
\name{as.data.frame.bcf_player}
\alias{as.data.frame.bcf_player}
\title{S3 as.data.frame method for player}
\usage{
\method{as.data.frame}{bcf_player}(x, ...)
}
\arguments{
\item{x}{S3 player object.}

\item{...}{Additional parameters.}
}
\value{
A data frame.
}
\description{
S3 as.data.frame method for player
}
\keyword{internal}
