% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coin_flip.R
\name{coin_flip_game}
\alias{coin_flip_game}
\title{Single coin flip game allowing multiple coins}
\usage{
coin_flip_game(thetas)
}
\arguments{
\item{thetas}{Vector of probabilities of success for each coin}
}
\value{
A vector of ranks for each probability in \code{thetas}
}
\description{
Each coin is given a probability of a success, independent of other coins.
Ties are dealt with by recursion, though I barely understand how I got this
to work. If all coins return 1 or all coins return 0 (an all-way tie), simply
repeat the flip; nothing has changed. If any coins come out ahead, partition
the different coins by 'rank' (rank 1 means a coin succeeded while others did
not) and recursively run the simulation until all ties have been broken and
each coin has been assigned a unique rank, 1 through \eqn{N}.
}
\examples{
example_game <- coin_flip_game(c(0.5, 0.5, 0.7))
}
\keyword{internal}
