% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player.R
\name{fit_beta_shape}
\alias{fit_beta_shape}
\title{Fit a Beta distribution to samples}
\usage{
fit_beta_shape(posterior, prior = list(shape1 = 1, shape2 = 1))
}
\arguments{
\item{posterior}{Samples from the distribution to fit.}

\item{prior}{Prior shape.}
}
\value{
A list with new Beta parameters.
}
\description{
Fit a Beta distribution to samples
}
\keyword{internal}
