% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player.R
\name{get_name}
\alias{get_name}
\title{Return a player's name}
\usage{
get_name(.player)
}
\arguments{
\item{.player}{A player object.}
}
\value{
A character vector.
}
\description{
Convenience function for calling \code{lapply()} against a list of players.
}
\keyword{internal}
