% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player.R
\name{get_uuid}
\alias{get_uuid}
\title{Return a player's UUID}
\usage{
get_uuid(.player)
}
\arguments{
\item{.player}{A player object.}
}
\value{
A character vector.
}
\description{
Convenience function for calling \code{lapply()} against a list of players.
}
\keyword{internal}
