% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game.R
\name{match_samples}
\alias{match_samples}
\title{Match samples corresponding to a result}
\usage{
match_samples(samples, result)
}
\arguments{
\item{samples}{A data frame of samples.}

\item{result}{A numeric vector describing a result.}
}
\value{
A logical vector.
}
\description{
Match samples corresponding to a result
}
\keyword{internal}
