% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game.R
\name{new_game}
\alias{new_game}
\title{Create a new game}
\usage{
new_game(players, result = NULL, samples = NULL, ...)
}
\arguments{
\item{players}{A list of players, each created with \code{\link{new_player}}}

\item{result}{The final ranking of players in the game. The winner is
assigned rank, second place receives rank 2, etc}

\item{samples}{Number of samples to draw from the join distribution
\eqn{(d*, \theta*)}}

\item{...}{Additional parameters to set in the game}
}
\value{
A \code{bcf_game} object
}
\description{
Create a new game
}
\keyword{internal}
