% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player.R
\name{new_player}
\alias{new_player}
\title{Create a new player}
\usage{
new_player(name, alpha = 1, beta = 1, ...)
}
\arguments{
\item{name}{Player name.}

\item{alpha}{Beta distribution shape parameter.}

\item{beta}{Beta distribution shape parameter.}

\item{...}{Other parameters assigned to the player.}
}
\value{
A \code{bcf_player} object.
}
\description{
Create a new player
}
\examples{
player1 <- new_player("Tom", alpha = 0.5, beta = 0.5)
}
