% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcf.R
\name{parbcf}
\alias{parbcf}
\title{Run BCF in parallel}
\usage{
parbcf(y, z, x_control, x_moderate = x_control, pihat, w = NULL,
  n_threads = max(RcppParallel::defaultNumThreads()/2, 1),
  random_seed = c(1, 235, 2134, 1236), nburn, nsim, nthin = 1,
  update_interval = 100, ntree_control = 200, sd_control = 2 * sd(y),
  base_control = 0.95, power_control = 2, ntree_moderate = 50,
  sd_moderate = sd(y), base_moderate = 0.25, power_moderate = 3,
  nu = 3, lambda = NULL, sigq = 0.9, sighat = NULL,
  include_pi = "control", use_muscale = TRUE, use_tauscale = TRUE,
  verbose = FALSE, nchains = 4)
}
\arguments{
\item{y}{Response variable}

\item{z}{Treatment variable}

\item{x_control}{Design matrix for the "prognostic" function mu(x)}

\item{x_moderate}{Design matrix for the covariate-dependent treatment effects tau(x)}

\item{pihat}{Length n estimates of propensity score}

\item{w}{An optional vector of weights. When present, BCF fits a model \eqn{y | x ~ N(f(x), \sigma^2 / w)}, where \eqn{f(x)} is the unknown function.}

\item{n_threads}{An optional integer of the number of threads to parallelize bcf operations on}

\item{random_seed}{A random seed passed to r's set.seed (vector as long as nchains)}

\item{nburn}{Number of burn-in MCMC iterations}

\item{nsim}{Number of MCMC iterations to save after burn-in}

\item{nthin}{Save every nthin'th MCMC iterate. The total number of MCMC iterations will be nsim*nthin + nburn.}

\item{update_interval}{Print status every update_interval MCMC iterations}

\item{ntree_control}{Number of trees in mu(x)}

\item{sd_control}{SD(mu(x)) marginally at any covariate value (or its prior median if use_muscale=TRUE)}

\item{base_control}{Base for tree prior on mu(x) trees (see details)}

\item{power_control}{Power for the tree prior on mu(x) trees}

\item{ntree_moderate}{Number of trees in tau(x)}

\item{sd_moderate}{SD(tau(x)) marginally at any covariate value (or its prior median if use_tauscale=TRUE)}

\item{base_moderate}{Base for tree prior on tau(x) trees (see details)}

\item{power_moderate}{Power for the tree prior on tau(x) trees (see details)}

\item{nu}{Degrees of freedom in the chisq prior on \eqn{sigma^2}}

\item{lambda}{Scale parameter in the chisq prior on \eqn{sigma^2}}

\item{sigq}{Calibration quantile for the chisq prior on \eqn{sigma^2}}

\item{sighat}{Calibration estimate for the chisq prior on \eqn{sigma^2}}

\item{include_pi}{Takes values "control", "moderate", "both" or "none". Whether to
include pihat in mu(x) ("control"), tau(x) ("moderate"), both or none. Values of "control"
or "both" are HIGHLY recommended with observational data.}

\item{use_muscale}{Use a half-Cauchy hyperprior on the scale of mu.}

\item{use_tauscale}{Use a half-Normal prior on the scale of tau.}

\item{verbose}{logical, whether to print log of MCMC iterations, defaults to FALSE.}

\item{nchains}{numeric, number of chains to launch in parallel (for convergence diagnostics). Should be less or equal to number of cores on the machine.}
}
\value{
A list with elements
\item{tau}{\code{nsim} by \code{n} matrix of posterior samples of individual treatment effects}
\item{mu}{\code{nsim} by \code{n} matrix of posterior samples of individual treatment effects}
\item{sigma}{Length \code{nsim} vector of posterior samples of sigma}
}
\description{
Run BCF in parallel
}
