% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player.R
\name{player_map_theta}
\alias{player_map_theta}
\title{Maximum A Posteriori win parameter}
\usage{
player_map_theta(player)
}
\arguments{
\item{player}{A \code{bcf_player} object}
}
\value{
A double
}
\description{
Return the MAP win parameter, the mode of the beta distribution
}
\keyword{internal}
