% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game.R
\name{plot.bcf_game}
\alias{plot.bcf_game}
\title{S3 plot method for a game}
\usage{
\method{plot}{bcf_game}(x, ...)
}
\arguments{
\item{x}{S3 player object.}

\item{...}{Additional parameters.}
}
\value{
A ggplot2 plot.
}
\description{
S3 plot method for a game
}
\keyword{internal}
