% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player.R
\name{plot.bcf_player}
\alias{plot.bcf_player}
\title{S3 plot method for a player}
\usage{
\method{plot}{bcf_player}(x, ...)
}
\arguments{
\item{x}{S3 player object.}

\item{...}{Additional parameters.}
}
\value{
A ggplot2 plot.
}
\description{
S3 plot method for a player
}
\examples{
\dontrun{
tom <- new_player("Tom", alpha = 2, beta = 2)
plot(tom)
}
}
\keyword{internal}
