% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game.R
\name{print.bcf_game}
\alias{print.bcf_game}
\title{S3 print method for game}
\usage{
\method{print}{bcf_game}(x, ...)
}
\arguments{
\item{x}{The object to be printed}

\item{...}{Additional parameters}
}
\value{
The \code{bcf_game} object, invisibly
}
\description{
S3 print method for game
}
\keyword{internal}
