% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player.R
\name{print.bcf_player}
\alias{print.bcf_player}
\title{S3 print method for player}
\usage{
\method{print}{bcf_player}(x, ...)
}
\arguments{
\item{x}{Object to be printed.}

\item{...}{Additional parameters.}
}
\value{
The \code{bcf_player} object, invisibly
}
\description{
S3 print method for player
}
\keyword{internal}
