% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coin_flip.R
\name{rep_coin_flip_game}
\alias{rep_coin_flip_game}
\title{Repeat many coin flip games, drawing from players' priors}
\usage{
rep_coin_flip_game(.players, .iters)
}
\arguments{
\item{.players}{A list of players.}

\item{.iters}{An integer number of iterations to run.}
}
\value{
A data frame of results, ordered following \code{.players}.
}
\description{
Repeat many coin flip games, drawing from players' priors
}
\keyword{internal}
