% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game.R
\name{summary.bcf_game}
\alias{summary.bcf_game}
\title{S3 summary method for a game}
\usage{
\method{summary}{bcf_game}(object, ...)
}
\arguments{
\item{object}{S3 game object.}

\item{...}{Additional parameters.}
}
\value{
A data frame.
}
\description{
S3 summary method for a game
}
\keyword{internal}
