% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player.R
\name{update_player}
\alias{update_player}
\title{Update a player object using a game result}
\usage{
update_player(player, game)
}
\arguments{
\item{player}{A player.}

\item{game}{A game.}
}
\value{
An updated \code{bcf_player} object.
}
\description{
Update a player object using a game result
}
\examples{
\dontrun{
players <- list(new_player("Tom"), new_player("Bob"))
sim_game <- abc_coin_flip_game(players)

players[[1]] <- update_player(players[[1]], sim_game)
players[[2]] <- update_player(players[[2]], sim_game)
}
}
