% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcf.R
\name{predict}
\alias{predict}
\title{Predict from Previously Fit Forests}
\usage{
predict(bcf_out, x_predict_control, x_predict_moderate, pi_pred, z_pred,
  save_tree_directory, n_chain_clusters = 2)
}
\arguments{
\item{bcf_out}{output from a BCF predict run}

\item{x_predict_control}{matrix of covariates for the "prognostic" function mu(x) for predictions (optional)}

\item{x_predict_moderate}{matrix of covariates for the covariate-dependent treatment effects tau(x) for predictions (optional)}

\item{pi_pred}{propensity score for prediction}

\item{z_pred}{Treatment variable for predictions (optional except if x_pre is not empty)}

\item{save_tree_directory}{directory where the trees have been saved}

\item{n_chain_clusters}{An optional integer of the number of clusters to run your MCMC chains on}
}
\description{
Predict from Previously Fit Forests
}
