% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcf.R
\name{predict}
\alias{predict}
\title{Predict from Previously Fit Forests}
\usage{
predict()
}
\arguments{
\item{y}{Response variable}

\item{z}{Treatment variable}
}
\description{
Predict from Previously Fit Forests
}
