% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcf.R
\name{summarise_bcf}
\alias{summarise_bcf}
\title{Print Summary Stats for a previous BCF run}
\usage{
summarise_bcf(bcf_out)
}
\arguments{
\item{bcf_out}{output from a BCF predict run}
}
\description{
Print Summary Stats for a previous BCF run
}
